remove hep_plotters.py
