+++ docs/documentation/documentation.rst
@@ -11,22 +11,16 @@ plotters.py
 .. automodule:: plothist.plotters
    :members:
 
-variable_registry.py
------------
-
-.. automodule:: plothist.variable_registry
-   :members:
-
-hep_plotters.py
----------------
-
-.. automodule:: plothist.hep_plotters
-   :members:
-
 comparison.py
 ---------------
 
 .. automodule:: plothist.comparison
+   :members:
+
+variable_registry.py
+-----------
+
+.. automodule:: plothist.variable_registry
    :members:
 
 plothist_style.py
